+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add http fields to `span.data` ([#1497](https://github.com/getsentry/sentry-dart/pull/1497))
   - Set `http.response.status_code`
   - Set `http.response_content_length`
+- Improve `SentryException#value`, remove stringified stack trace ([##1470](https://github.com/getsentry/sentry-dart/pull/#1470))
 
 ## 7.6.3